+++ CHANGES.md
@@ -1,8 +1,8 @@
 **Added:**
-* untested backup for installer
+* 
 
 **Changed:**
-* more of the gui for installer window implemented
+* finished (?) header for installer window
 
 **Fixed:**
 * none